bug fixes/functions file added
