+++ Kod.js
@@ -1,5 +1,6 @@
 const Plugin = require("./Plugin.js");
 const { MessageEmbed } = require("discord.js");
+const { getMember } = require("./functions.js");
 var app = null;
 
 class Kod extends Plugin {
@@ -8,27 +9,6 @@ class Kod extends Plugin {
     commands = {
         'kod': function(message, args) {
 
-            function getMember(message, toFind = '') {
-                toFind = toFind.toLowerCase;
-
-                let target = message.guild.members.cache.get(toFind);
-
-                if (!target && message.mentions.members)
-                    target = message.mentions.members.first();
-
-                if (!target && toFind) {
-                    target = message.guild.members.cache.find(member => {
-                        return member.displayName.toLowerCase().includes(toFind) ||
-                            member.user.tag.toLowerCase().includes(toFind);
-                    });
-                }
-
-                if (!target)
-                    target = message.member;
-
-                return target;
-            }
-
             const member = getMember(message, message.author);
 
             if (message.deletable) {
@@ -36,7 +16,9 @@ class Kod extends Plugin {
             }
 
             let language = args[1];
-            console.log(language);
+            if (!language) return message.reply("``Använding: -kod <språk> <kod>``, Botten supportar följande språk => `js` `html` `css` `cs` `cpp`");
+
+
             let koden = args.join(" ").slice(language.length + 6);
 
             if (koden.length >= 1023) return message.reply("Det går tyvärr inte att klistra in så lång kod än :/");
