+++ package.json
@@ -1,25 +1,25 @@
 {
-  "name": "lugbot",
-  "version": "1.0.0",
-  "description": "",
-  "main": "lugbot.js",
-  "dependencies": {
-    "ascii-table": "0.0.9",
-    "cheerio": "^1.0.0-rc.3",
-    "common-tags": "^1.8.0",
-    "discord.js": "^12.5.0",
-    "dotenv": "^8.2.0",
-    "fs": "0.0.1-security",
-    "load-svg-file": "^1.0.8",
-    "random": "^2.2.0",
-    "request": "^2.88.2",
-    "request-promise": "^4.2.6",
-    "svg-to-img": "^2.0.9"
-  },
-  "devDependencies": {},
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "Emil Österlund",
-  "license": "GPL-3.0-or-later"
+    "name": "lugbot",
+    "version": "1.0.0",
+    "description": "",
+    "main": "main.js",
+    "dependencies": {
+        "ascii-table": "0.0.9",
+        "cheerio": "^1.0.0-rc.3",
+        "common-tags": "^1.8.0",
+        "discord.js": "^12.5.0",
+        "dotenv": "^8.2.0",
+        "fs": "0.0.1-security",
+        "load-svg-file": "^1.0.8",
+        "random": "^2.2.0",
+        "request": "^2.88.2",
+        "request-promise": "^4.2.6",
+        "svg-to-img": "^2.0.9"
+    },
+    "devDependencies": {},
+    "scripts": {
+        "test": "echo \"Error: no test specified\" && exit 1"
+    },
+    "author": "Emil Österlund",
+    "license": "GPL-3.0-or-later"
 }
